requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
jinja2==3.1.5
aiofiles==24.1.0
sqlalchemy[asyncio]==2.0.36
aiosqlite==0.20.0
python-jose[cryptography]==3.3.0
bcrypt==4.2.1
aiosmtplib==3.0.2
python-dotenv==1.0.1
email-validator==2.2.0
apscheduler==3.10.4
httpx==0.28.1